backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
sqlalchemy==2.0.20
python-multipart==0.0.6
aiofiles==23.2.1
watchdog==3.0.0
python-dotenv==1.0.0 